[#45] Chore: Github Service package structure segmentation.
