+++ src/main/java/soon/capstone/domain/team/service/team/TeamJoinService.java
@@ -10,7 +10,7 @@ import soon.capstone.domain.team.repository.TeamRepository;
 import soon.capstone.domain.teammember.entity.TeamMember;
 import soon.capstone.domain.teammember.repository.TeamMemberRepository;
 import soon.capstone.global.exception.teammember.AlreadyTeamMemberException;
-import soon.capstone.infrastructure.github.service.GithubOrganizationService;
+import soon.capstone.infrastructure.github.service.organization.GithubOrganizationService;
 import soon.capstone.infrastructure.redis.invitation.entity.InvitationCode;
 import soon.capstone.infrastructure.redis.invitation.repository.InvitationCodeRepository;
