+++ src/main/java/soon/capstone/infrastructure/github/service/project/GithubLinkProjectToRepositoryService.java
@@ -1,4 +1,4 @@
-package soon.capstone.infrastructure.github.service;
+package soon.capstone.infrastructure.github.service.project;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
